+++ docs/changelog.rst
@@ -12,6 +12,12 @@ Changed
 * Don't unregister log handler explicitly in the destructor, this is now
   handled by the logging library.
 
+Fixed
+~~~~~
+
+* Make registration of shells safe during static initialization.
+* Make use of the built-in logger instance safe during static initialization.
+
 0.7.3_ |--| 2019-09-22
 ----------------------